rag-codegen/requirements.txt
huggingface_hub
sentence-transformers
faiss-cpu
python-dotenv
numpy